Fixed a type bug in FireflyAPIDataClass subclasses
The Firefly API is returning a JSON and Python json package is reading
everything it can as string. It means that integers such as the ones
that can be found for instance IDs were strings when created from the
API. This is no longer true, and now integers, floats and datetimes
are correctly parsed.